# laravel/openapi.yaml
openapi: 3.0.0
info:
  title: 'API Documentation'
  contact:
    email: info@mail.com
  version: '0.1'
servers:
  -
    url: 'https://localhost/api/'
    description: Server
paths:
  /auto:
    get:
      tags:
        - Auto
      operationId: 245ac8a371c45c1c483c49c96116fd5f
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/AutoResource'
        '401':
          description: Unauthorized
        '404':
          description: 'Not Found'
    post:
      tags:
        - Auto
      operationId: cae1ab0d913d32c1e71066d832dc1af9
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AutoRequest'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AutoResource'
        '401':
          description: Unauthorized
        '404':
          description: 'Not Found'
  '/auto/{id}':
    delete:
      tags:
        - Auto
      operationId: 1dc1c9060f0e18e3dee7db082bd526f8
      parameters:
        -
          name: id
          in: path
          description: 'ID of auto'
          required: true
          schema:
            type: integer
      responses:
        '201':
          description: Ok
        '401':
          description: Unauthorized
        '404':
          description: 'Not Found'
    patch:
      tags:
        - Auto
      operationId: beea1edf8a9fdfb7d9547189da17ebdb
      parameters:
        -
          name: id
          in: path
          description: 'ID of auto'
          required: true
          schema:
            type: integer
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AutoRequest'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AutoResource'
        '401':
          description: Unauthorized
        '404':
          description: 'Not Found'
  /auto_mark:
    get:
      tags:
        - Auto
      operationId: a3ecdf9e756384de939ade4aaf2b87d7
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/AutoMarkResource'
        '401':
          description: Unauthorized
        '404':
          description: 'Not Found'
    post:
      tags:
        - Auto
      operationId: f97e01c91502559c0916735cb4865325
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AutoMarkRequest'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AutoMarkResource'
        '401':
          description: Unauthorized
        '404':
          description: 'Not Found'
  '/auto_mark/{id}':
    delete:
      tags:
        - Auto
      operationId: d453fd4b9ec596e886186dd189ca8247
      parameters:
        -
          name: id
          in: path
          description: 'ID of auto mark'
          required: true
          schema:
            type: integer
      responses:
        '201':
          description: Ok
        '401':
          description: Unauthorized
        '404':
          description: 'Not Found'
    patch:
      tags:
        - Auto
      operationId: 84f63c3323a5305e2404d3ff95072f24
      parameters:
        -
          name: id
          in: path
          description: 'ID of auto mark'
          required: true
          schema:
            type: integer
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AutoMarkRequest'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AutoMarkResource'
        '401':
          description: Unauthorized
        '404':
          description: 'Not Found'
...
components:
  schemas:
    AutoMarkRequest:
      required:
        - name
      properties:
        name:
          $ref: '#/components/schemas/AutoMark/properties/name'
      type: object
...
    AutoRequest:
      required:
        - year
        - mileage
        - color
        - model_id
        - mark_id
      properties:
        year:
          $ref: '#/components/schemas/Auto/properties/year'
        mileage:
          $ref: '#/components/schemas/Auto/properties/mileage'
        color:
          $ref: '#/components/schemas/Auto/properties/color'
        model_id:
          $ref: '#/components/schemas/Auto/properties/auto_model_id'
        mark_id:
          $ref: '#/components/schemas/Auto/properties/auto_mark_id'
        user_id:
          $ref: '#/components/schemas/Auto/properties/user_id'
      type: object
...
    AutoMark:
      properties:
        id:
          description: ID
          type: integer
          example: '1'
        name:
          description: 'AutoMark name'
          type: string
          example: BMW
        created_at:
          description: 'Created at'
          type: string
          format: date-time
          example: '2025-01-01 00:00:01'
        updated_at:
          description: 'Updated at'
          type: string
          format: date-time
          example: '2025-01-01 00:00:01'
      type: object
    AutoModel:
      properties:
        id:
          description: ID
          type: integer
          example: '1'
        name:
          description: 'AutoModel name'
          type: string
          example: M6
        created_at:
          description: 'Created at'
          type: string
          format: date-time
          example: '2025-01-01 00:00:01'
        updated_at:
          description: 'Updated at'
          type: string
          format: date-time
          example: '2025-01-01 00:00:01'
      type: object
    AutoMarkResource:
      properties:
        id:
          $ref: '#/components/schemas/AutoMark/properties/id'
        name:
          $ref: '#/components/schemas/AutoMark/properties/name'
        created_at:
          $ref: '#/components/schemas/AutoMark/properties/created_at'
        updated_at:
          $ref: '#/components/schemas/AutoMark/properties/updated_at'
      type: object
...
    AutoResource:
      properties:
        data:
          properties:
            id:
              $ref: '#/components/schemas/Auto/properties/id'
            year:
              $ref: '#/components/schemas/Auto/properties/year'
            mileage:
              $ref: '#/components/schemas/Auto/properties/mileage'
            color:
              $ref: '#/components/schemas/Auto/properties/color'
            model:
              $ref: '#/components/schemas/AutoModel'
            mark:
              $ref: '#/components/schemas/AutoMark'
            created_at:
              $ref: '#/components/schemas/Auto/properties/created_at'
            updated_at:
              $ref: '#/components/schemas/Auto/properties/updated_at'
          type: object
      type: object
...